src/requirements.txt
anyio==3.6.2
asttokens==2.1.0
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
branca==0.5.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Cython==0.29.28
debugpy==1.6.3
decorator==5.1.1
dnspython==2.2.1
entrypoints==0.4
exceptiongroup==1.0.1
executing==1.2.0
fastapi==0.86.0
fasttext-win==0.8.3
future==0.18.2
gensim==4.2.0
googlemaps==4.6.0
h11==0.14.0
idna==3.4
ipykernel==6.17.0
ipyleaflet==0.17.2
ipython==8.6.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jupyter-client==7.4.4
jupyter-core==4.11.2
jupyterlab-widgets==3.0.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.4
outcome==1.2.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
pip==21.1.2
prompt-toolkit==3.0.32
psutil==5.9.3
pure-eval==0.2.2
pybind11==2.10.1
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pymongo==4.3.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pywin32==305
pyzmq==24.0.1
requests==2.28.1
scikit-learn==1.2.0
scipy==1.9.3
selenium==4.6.0
setuptools==57.0.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.0
starlette==0.20.4
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
traittypes==0.2.1
trio-websocket==0.9.2
trio==0.22.0
typing-extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
wcwidth==0.2.5
webdriver-manager==3.8.4
wheel==0.36.2
widgetsnbextension==4.0.3
wsproto==1.2.0
xyzservices==2022.9.0

